Moved logging config model into utilities/logging dir
- I think it makes more sense here since it's not really app specific
